+++ apps/multiplatform/common/src/desktopMain/kotlin/chat/simplex/common/DesktopApp.kt
@@ -23,6 +23,7 @@ import chat.simplex.res.MR
 import dev.icerock.moko.resources.compose.painterResource
 import dev.icerock.moko.resources.compose.stringResource
 import kotlinx.coroutines.*
+import java.awt.Frame
 import java.awt.event.WindowEvent
 import java.awt.event.WindowFocusListener
 import java.io.File
@@ -251,6 +252,17 @@ private fun ApplicationScope.handleCloseRequest(closedByError: MutableState<Bool
       pref.set(CloseBehavior.Quit)
       exitApplication()
     }
+  }
+}
+
+fun showWindow() {
+  simplexWindowState.windowVisible.value = true
+  simplexWindowState.window?.apply {
+    // Clear ICONIFIED so a minimized window un-minimizes; preserves MAXIMIZED_BOTH
+    // when set. toFront() alone does not un-minimize on any AWT platform.
+    extendedState = extendedState and Frame.ICONIFIED.inv()
+    toFront()
+    requestFocus()
   }
 }
